Cleaned up AboutInfo component
